Force macro compilation by importing all files in system and component subdirs
Fun problem: importing via the systems.json file only actually imports
files that were already compiled. If a system isn't imported by anything
else, there's nothing to include it in the json file, and then the macro
doesn't run. Force-import everything first to get the metadata, then
generate the update proc.
This of course means that including the filename itself in the metadata
is currently pointless. Perhaps future-me will find a use for it? Or
delete it eventually.